+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 NOWnews-api change logs
 
 ## [Unreleased]
+### Added
+- cronjob加上軍聞社新聞匯入 @appleoxxo
 
 ## 1.0.30 - 2017-09-04
 ### Fixed
@@ -18,7 +20,6 @@ NOWnews-api change logs
 - 暫時把鉅亨網新聞匯入註解掉 @appleoxxo
 
 ### Added
-- cronjob加上軍聞社新聞匯入 @appleoxxo
 - 刪除新聞的時候，也要有 log (因應這個功能，所以在 `newLog` 這個 schema 加入新的欄位 `newsTrashed`) @SimonSun
 - cronjob加上中央社圖片匯入 @appleoxxo