+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 NOWnews-api change logs
 
 ## [Unreleased]
+- 加入 imagelab 縮圖網址與在 Image mode 加入 sizeFormat 虛擬欄位 @SimonSun
 
 ## 1.0.30 - 2017-09-04
 ### Fixed