+++ CHANGELOG.md
@@ -6,6 +6,9 @@ NOWnews-api change logs
 ### Fixed
 - 修正 google trace engine 造成 CPU 與 Memory 消耗過大問題 @SimonSun
 
+### Added
+- 加入 imagelab 縮圖網址與在 Image mode 加入 sizeFormat 虛擬欄位 @SimonSun
+
 ## 1.0.30 - 2017-09-04
 ### Fixed
 - 修正 中央社匯入圖片有IP限制 外網看不到的問題 @appleoxxo